Move code to hook
